Updated the logger for Sentinel version .12.  Now will automatically log location of event. Implemented authentication methods. Added restraints to the Username sent to the server. Added more complete a disposal method for both client and server. All tests pass, but currently can not be run in parallel since they are all binding to the same address.

+++ DtronixJsonRpcTests/TestClientActions.cs
@@ -16,7 +16,7 @@ namespace DtronixJsonRpcTests {
 
 
 		public class TestClientActionTestArgs : JsonRpcActionArgs {
-			public int RandomInt { get; set; }
+			public long RandomLong { get; set; }
 		}
 
 		[ActionMethod(JsonRpcSource.Client)]
